backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
starlette
deepgram-sdk
aiofiles
python-multipart
